#10 refactor : Comment
merge시에 변경점 적용
